app/PolicyAgent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PolicyAgent"
version = "0.1.0"
description = "Insurance policy Q&A agent on AWS Bedrock AgentCore"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "strands-agents >= 1.13.0",
    "PyPDF2 >= 3.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
